Migrated /stats/queries route code to stats.py (no need to clutter w. sep. src file)
